Update
The GET, POST and DELETE request work. I wrote the function for the PUT request, but still could not find out how to call it from an eventListener after changing the of the todo's.
I also tried another trash-icon from fontawesome.com, and now I get to see the trashcan but if gives an error when firing the deleteTodoApi() function. If I use the img tag works fine, but with the button tag it gives the error.